Increase size of Number to be compliant with Lox spec

+++ src/scanner.rs
@@ -124,7 +124,7 @@ impl<'a> Scanner<'a> {
             }
         }
         let result = self.src.inner()[self.start..self.at]
-            .parse::<f32>()
+            .parse::<f64>()
             .expect("parsing of float an not fail");
         TokenType::Number(result)
     }
